requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
python-dotenv==1.0.0
Werkzeug==2.2.3

SQLAlchemy~=2.0.15